nvme: add separate handling for platform IRQ
On a platform device, there is no PCI config space, and thus
no way to configure MSI interrupts, so every interrupt has
to go through through the same line. This also assumes that
there is a single shared queue.

Guard all PCI interrupt specific function calls with a check
for the PCI device, and provide an alternative where this
is needed.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>